Launch SuperSU from Settings
If SuperSU is installed, show an entry in Settings below
"Developer options", which launches SuperSU when tapped.
At boot time SuperSU will detect if com.android.settings
contains the resources from this patch, and if so it
will hide itself from the launcher. This was implemented
in v1.85, so we only show the entry if the installed
SuperSU version is at least that.

PS3 - Added material green icons for SU

Change-Id: I2af95a1a4396101ff87448cd0ae20addf235af52

Conflicts:
	res/values/strings.xml

Change-Id: Id68b09c0113ac5f71d47223ce28d04d5f5d72bd6