Settings: Update AppNotificationSettings if user resumes
Right now if a user resumes this activty (eg from recents) and did change in the
mean time settings which effects the visibility of preferences in AppNotificationSettings
(eg were removed) it does not get updated.

Fix it with simply reloading the content on onResume and decide to show the entries
on the current user settings situation.