Settings: Add Advanced reboot (1 of 2)
This commit adds a setting in Development settings for including options
in the power menu for rebooting into recovery or bootloader, defauled to
off.

When enabled, the Advanced reboot options will only be available once the
device is unlocked.

Modified from this CM commit= 28ad8e332631198a987d55eed13440a5332ec19a
To work with AOSP

Change-Id: I17ccc3d9722c88f5500cd9c8c1873fd7cef631a2